250108 mysite03 - guestbook transaction

+++ mysite03/src/main/java/mysite/service/GuestbookService.java
@@ -3,28 +3,49 @@ package mysite.service;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import mysite.repository.GuestbookLogRepository;
 import mysite.repository.GuestbookRepository;
 import mysite.vo.GuestbookVo;
 
 @Service
 public class GuestbookService {
 	
-	private GuestbookRepository guestbookRepository;
+	private final GuestbookRepository guestbookRepository;
+	private final GuestbookLogRepository guestbookLogRepository;
 	
-	public GuestbookService(GuestbookRepository guestbookRepository) {
+	public GuestbookService(GuestbookRepository guestbookRepository, GuestbookLogRepository guestbookLogRepository) {
 		this.guestbookRepository = guestbookRepository;
+		this.guestbookLogRepository = guestbookLogRepository;
 	}
 	
 	public List<GuestbookVo> getContentsList(){
 		return guestbookRepository.findAll();
 	}
 	
+	@Transactional
 	public void deleteContents(Long id, String password) {
-		guestbookRepository.deleteByIdAndPassword(id, password);
+		GuestbookVo vo = guestbookRepository.findById(id);
+		if(vo == null) {
+			return;
+		}
+		
+		int count = guestbookRepository.deleteByIdAndPassword(id, password);
+		//password가 잘못되었을 경우 안지워질 수 있다
+		if(count == 1) {
+			guestbookLogRepository.updateByRegDate(vo.getRegDate());
+		}
 	}
 	
+	@Transactional
 	public void addContents(GuestbookVo vo) {
+		System.out.println(vo);
+		int count = guestbookLogRepository.update();
+		if(count==0) {
+			guestbookLogRepository.insert();
+		}
+		
 		guestbookRepository.insert(vo);
 	}
 }
